+++ client/src/dependencyMgr.ts
@@ -7,22 +7,23 @@
  * the LICENSE.txt file in the root directory of this source tree.
  */
 
-import * as Q from "q";
+import * as Q from 'q';
 
-import { IDependency } from "./interfaces";
-import { Dependency } from "./dependency";
-import { Version } from "./version";
-import { workspace } from "vscode";
-
+import { IDependency } from './interfaces';
+import { Dependency } from './dependency';
+import { Version } from './version';
+import { workspace } from 'vscode';
 
 export class DependencyMgr {
     static _dependencies: Q.Promise<IDependency[]> = [];
     constructor() {
         const config = workspace.getConfiguration('ibm-q-studio');
-        const dependList = config.get("python.dependencies");
+        const dependList = config.get('python.dependencies');
 
-        Object.keys(dependList).forEach(function (key) {
-            DependencyMgr._dependencies.push(new Dependency(key.toString(), Version.fromString(dependList[key].toString())));
+        Object.keys(dependList).forEach(function(key) {
+            DependencyMgr._dependencies.push(
+                new Dependency(key.toString(), Version.fromString(dependList[key].toString()))
+            );
         });
     }